POLIO-1217: add Audit to reasons for delay
