+++ iaso/api/microplanning.py
@@ -6,6 +6,7 @@ from rest_framework.decorators import action
 from rest_framework.fields import Field
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
+from hat.audit.audit_mixin import AuditMixin
 from iaso.permissions import ReadOnly
 
 from hat.audit.models import Modification
@@ -187,53 +188,6 @@ class TeamAncestorFilterBackend(filters.BaseFilterBackend):
             queryset = queryset.filter(path__descendants=ancestor.path)
 
         return queryset
-
-
-class AuditMixin:
-    audit_serializer: serializers.ModelSerializer
-
-    def perform_create(self, serializer):
-        # noinspection PyUnresolvedReferences
-        super().perform_create(serializer)
-        instance = serializer.instance
-
-        serialized = [self.audit_serializer(instance).data]
-        Modification.objects.create(
-            user=self.request.user,
-            past_value=[],
-            new_value=serialized,
-            content_object=instance,
-            source="API " + self.request.method + self.request.path,
-        )
-
-    def perform_update(self, serializer):
-        instance = serializer.instance
-        old_value = [self.audit_serializer(instance).data]
-        # noinspection PyUnresolvedReferences
-        super().perform_update(serializer)
-        instance = serializer.instance
-        new_value = [self.audit_serializer(instance).data]
-        Modification.objects.create(
-            user=self.request.user,
-            past_value=old_value,
-            new_value=new_value,
-            content_object=instance,
-            source="API " + self.request.method + self.request.path,
-        )
-
-    def perform_destroy(self, instance):
-        old_value = [self.audit_serializer(instance).data]
-        # noinspection PyUnresolvedReferences
-        super().perform_destroy(instance)
-        # for soft delete, we still have an existing instance
-        new_value = [self.audit_serializer(instance).data]
-        Modification.objects.create(
-            user=self.request.user,
-            past_value=old_value,
-            new_value=new_value,
-            content_object=instance,
-            source=f"API {self.request.method} {self.request.path}",
-        )
 
 
 class TeamViewSet(AuditMixin, ModelViewSet):
